requirements.txt
pandas
scikit-learn
streamlit
requests
Pillow
